merkle-tree-lib/pyproject.toml
[tool.poetry]
name = "merkle-tree-lib"
version = "0.1.4"
description = ""
authors = ["Gabriel Fior <gabriel.fior@outlook.com>"]
readme = "README.md"
packages = [{include = "merkle_tree_lib"}]

[tool.poetry.dependencies]
python = "^3.9"
pysha3 = "^1.0.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
